python/pyproject.toml
[tool.poetry]
name = "rightsizing-analysis-server"
version = "0.1.0"
description = "rightsizing analysis server"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.42.0"
pandas = "^1.3.5"
pystan = "2.19.1.1"
fbprophet = "^0.7.1"
numpy = "^1.22.2"
google = "^3.0.0"
protobuf = "^3.19.4"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.42.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
